修复NGP在多线程环境下更新着色器变量无效的bug，添加水印滤镜、VibranceFilter等

+++ library/src/main/java/com/ben/android/library/filter/ZoomBlurFilter.java
@@ -21,7 +21,7 @@ public class ZoomBlurFilter extends NativeFilter {
     public ZoomBlurFilter(PointF pointF) {
         this(null, 0);
     }
-     public ZoomBlurFilter(PointF pointF,int blurSize) {
+     public ZoomBlurFilter(PointF pointF,float blurSize) {
         super(FilterType.ZOOM_BLUR.ordinal());
         if (pointF != null) {
             this.pointXF = pointF.x;
